+++ module/search/src/main/kotlin/io/github/siyual_park/search/pagination/OffsetPagePresenter.kt
@@ -1,6 +1,7 @@
 package io.github.siyual_park.search.pagination
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import io.github.siyual_park.view.Presenter
 import kotlinx.coroutines.reactor.awaitSingleOrNull
 import org.springframework.http.HttpHeaders
 import org.springframework.http.MediaType
