+++ module/view/.gitignore
@@ -1,0 +1,7 @@
+/.gradle
+/.idea
+/out
+/build
+*.iml
+*.ipr
+*.iws
